Added Reducer for Login Actions
Login => initiates the login action which is picked by the saga
middleware to perform api call to access_token and receive token.

LoginSuccess => Called when Login is Successful, adds access token to
the store and sets the isLoggedIn to true

LoginFailure => Called on failure and dispatches error message to the
store to be displayed in the toast

moved Login api call from the login page to the LoginSaga to be called
on Login action dispatch

Removed demo reducer and its related pages and files

Added redux-persist package to the application. To be implemented soon
after the logout reducer is added (most probably as a root reducer to
clear the entire store on logout)

TODO:
1. UserReducer and UserSaga => shifting getUser from Login page into UserSaga
2. Adding RefreshToken and TokenExpiry into Login
3. Configuring Redux-persist for maintaining state in local storage
4. Removing react-cookie from the application