pages/users: Create users
Add a button to open a modal with no userId set, so the modal action will now create a user. Toasts with success and error messages will be triggered on success/error respectively. However, need to close the modal once user is successfully created/updated - can't figure out how to do this.

Also removed some extra unused space from the dashboard, as well as add header tags to both the users page and the modal.

Signed-off-by: ksdfg <ksdfg123@gmail.com>