pages/users: use jpg image instead of png
Firefox seems to be changing the colour of png images slightly, so the background colour that should be #282d33 comes out as #2b2f35. Not a huge difference in colours, but it was visible to me and was frustrating. Hopefully I won't end up wasting too much time on useless issues like this, but it was getting hard to work when my base theme wasn't matching the logo pics I had exported from figma - I just couldn't stop thinking about it.

Now I promise to stop fixing stuff that isn't important and work on the mvp

Signed-off-by: ksdfg <ksdfg123@gmail.com>