pages/users/components/UserModal: add checks before submit
For user update, make sure that there is atleast one property set in submit object, and that not all values set in submit are the same as the selectedUser i.e. no real change has been made to the user.

For user create, make sure submit has the keys email, type and password. Also make sure that none of those values are empty or null.

Signed-off-by: ksdfg <ksdfg123@gmail.com>