pages/users: set minimum height to 80vh
This reduces the awkward space between header and content. Since the content earlier was limited to 70vh, and our app centers everything by default, we were essentially left with empty space at the top and bottom, since the content was too small to fill it all up. Hence increasing minHeight of the content made it fill up the empty space while still being in screen boundries

Signed-off-by: ksdfg <ksdfg123@gmail.com>