redux: reset state of each reducer manually on logout
Directly setting `state = undefined` on logout was causing the _persist key to completely disappear from the state. In this case, it seemed to be better to reset each reducer manually instead of reducing the entire state in one go, and as such preserving the _persist reducer (which honestly should not be reset, since it doesn't store data related to whether a user is logged in or not, just whether the state has been rehydrated or not).

Also removed useless LOGOUT_SUCCESS and CLEAR_LOGIN_ERRORS actions.

Signed-off-by: ksdfg <ksdfg123@gmail.com>