pages/homepage: add ability to upload images for profile pictures
Display the uploaded image, and store the base64 string to submit. Will tweak this to match what is expected in the backend later.

Also removed the header title since it was taking too much space, and added an IconButton for uploading image. Having the image as a button is probably unclear

Signed-off-by: ksdfg <ksdfg123@gmail.com>