pages/dashboard: Handle errors for all modules
Since there are multiple modules, with many pages that are accessible via dashboard potentially handling multiple modules, there is a risk for the same error to trigger multiple toasts if the errors were handled separately in each page.
Since the error messages are set per module, if we check for all error messages and render the respective error toasts in the root component itself (which is the dashboard component in this case) we can avoid the multiple toast situation with ease.
This also allows us to not have the headache of not having to set toast display for error messages in each child component of dashboard. However, in the future, if we need to make sure that a component does not trigger a toast for a certain module's error message, that can be done via a map of booleans set as a state variable in the dashboard component itself, with the child component switching the value for the specific module (essentially switching error toasts on or off).

Signed-off-by: ksdfg <ksdfg123@gmail.com>