pages/users: Show school name instead of school ID in List of Users
Instead of just passing `this.props.users` i.e. fetched list of users from backend to DataPage component for rendering rows in DataGrid, use it to initialise a state variable `users`. Once we have fetched list of all schools, replace the `school` parameter of each object in `users` (which is initially set to school ID) to the name of the respective school.

Signed-off-by: ksdfg <ksdfg123@gmail.com>