Convert profile_picture property of all users to be set in redux to it's respective URL
Whenever we perform any CRUD operation with user(s), always set the `user.profile_picture` to `${process.env.REACT_APP_BACKEND_URL}/profile_pictures/${user.profile_picture}` even if it is null.
The reason for doing so is that for Avatar to set the first letter of the alt text as the Displayed picture, the `src` link needs to be broken. It does not work if `null` or `undefined` is passed to `src` - in that case it sets a generic user icon.
Since `${process.env.REACT_APP_BACKEND_URL}/profile_pictures/null` is always going to be a broken link, this ensures that those without a profile picture get the first letter of their name/email on their temp profile picture.

pages/users & pages/homepage: set src of Avatar to the `profile_picture` property of user.

Signed-off-by: ksdfg <ksdfg123@gmail.com>