services/http-client: Make an `APICall` function to make calls to API endpoints with auth header.
Essentially a separate function so that we don't have to fetch token and set header each time. Since this was a very common repetition, thought it would be wise to simply abstract it into a function and call it everywhere.

Also modified school, user and year to use `APICall` instead of `httpClient`. auth still uses `httpClient` since no auth header is needed for login.

Signed-off-by: ksdfg <ksdfg123@gmail.com>