components/SchoolSelect: Create a reusable component for a select input field with all the schools as options.
Since any object with school as a foreign key will have to use a select input field to select school, might as well abstract it into a reusable component. As of now the two uses seem to be in UserModal and the upcoming YearModal.

Also replaced the existing TextField in UserModal with SchoolSelect