services/http-client: refresh tokens past expiry
If current date is equal to or greater than expiry, then try to refresh the tokens. On error during token refresh, logout and throw a custom error telling user to login again.

There is an edge case where the token expires just after the API call is made, but the probability of that happening is slim enough that we can afford to ignore it in our code. Adding or subtracting time to get dates can be tricky in JS, and it doesn't seem to be worth using a separate library (MomentJS) just for this highly improbable edge case.